refactor: moved alert and error msg to main page
